+++ src/admin/ajax/ajax-pin-task.php
@@ -25,7 +25,7 @@ try {
     isset( $_POST['task_link'] )
     && isset( $_POST['post_id'] )
     && isset( $_POST['nonce'] )
-    && wp_verify_nonce( $_POST['nonce'], 'ptc_completionist_pin_task' ) !== FALSE//phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+    && wp_verify_nonce( $_POST['nonce'], 'ptc_completionist' ) !== FALSE//phpcs:ignore WordPress.Security.ValidatedSanitizedInput
     && Asana_Interface::has_connected_asana()
   ) {
 
@@ -61,6 +61,7 @@ try {
     if ( $did_pin_task === FALSE ) {
       throw new \Exception( "Failed to pin the existing task to post $the_post_id.", 409 );
     }
+    // @TODO - Else, update the cache on success.
 
     $res['status'] = 'success';
     $res['code'] = 200;