+++ src/components/task/TaskList.jsx
@@ -1,0 +1,6 @@
+import TaskRow from './TaskRow.jsx';
+
+export default function TaskList({tasks}) {
+	const renderedTasks = tasks.map( t => <TaskRow key={t.gid} task={t} />);
+	return <div className="ptc-TaskList">{renderedTasks}</div>;
+}