+++ src/includes/class-asana-interface.php
@@ -29,6 +29,15 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 	class Asana_Interface {
 
 		/**
+		 * The ?opt_fields csv for Asana API requests.
+		 *
+		 * @since 3.1.0
+		 *
+		 * @var string TASK_OPT_FIELDS
+		 */
+		public const TASK_OPT_FIELDS = 'name,completed,notes,due_on,assignee,workspace,tags';
+
+		/**
 		 * The currently authenticated WordPress user's ID.
 		 *
 		 * @since 1.1.0
@@ -134,6 +143,10 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		 * or requests could fail due to request limits or server issues.
 		 */
 		private static function maybe_load_client( int $user_id = 0 ) : \Asana\Client {
+
+			if ( 0 === $user_id ) {
+				$user_id = get_current_user_id();
+			}
 
 			$asana_personal_access_token = Options::get( Options::ASANA_PAT, $user_id );
 			if (
@@ -541,11 +554,11 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		 * Attempts to retrieve task data. Providing the post id of the provided
 		 * pinned task gid will also attempt data self-healing.
 		 *
+		 * @since 3.1.0 Marked $opt_fields param as deprecated.
 		 * @since 1.0.0
 		 *
 		 * @param string $task_gid The gid of the task to retrieve.
-		 * @param string $opt_fields A csv of task fields to retrieve, excluding
-		 * 'workspace' and 'tags', which are appended for data healing.
+		 * @param string $opt_fields_deprecated Deprecated.
 		 * @param int $post_id Optional. The post ID on which the task belongs to
 		 * attempt self-healing on certain error responses. Default 0 to take no
 		 * action on failure.
@@ -554,12 +567,17 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		 * @throws \Exception The Asana client may not be authenticated or the API
 		 * request may fail. Additional custom exceptions are:
 		 * * 400: Invalid task gid - The provided task gid is invalid.
-		 * * 410: Unpinned task - The API returned 404, so the task was unpinned.
-		 * * 410: Unpinned Foreign Task - The task does not belong to the assigned
-		 * workspace or have the site tag, so it was unpinned.
-		 * * 0: Failed to get task data - This is presumably unreachable.
+		 * * 410: Invalid task - The task is no longer available or relevent.
 		 */
-		public static function maybe_get_task_data( string $task_gid, string $opt_fields, int $post_id = 0 ) : \stdClass {
+		public static function maybe_get_task_data( string $task_gid, string $opt_fields_deprecated = '', int $post_id = 0 ) : \stdClass {
+
+			if ( ! empty( $opt_fields_deprecated ) ) {
+				_deprecated_argument(
+					__FUNCTION__,
+					'3.1.0',
+					'$opt_fields is now a member constant, ' . __CLASS__ . '::TASK_OPT_FIELDS'
+				);
+			}
 
 			$task_gid = Options::sanitize( 'gid', $task_gid );
 			if ( empty( $task_gid ) ) {
@@ -569,7 +587,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 			try {
 
 				$asana = self::get_client();
-				$task = $asana->tasks->findById( $task_gid, [ 'opt_fields' => $opt_fields . ',workspace,tags' ] );
+				$task = $asana->tasks->findById( $task_gid, [ 'opt_fields' => self::TASK_OPT_FIELDS ] );
 
 				if (
 					isset( $task->workspace->gid )
@@ -626,7 +644,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 				throw $e;
 			}
 
-			throw new \Exception( 'Failed to get task data', 0 );
+			throw new \Exception( 'Failed to get task data.', 0 );
 		}
 
 		/**
@@ -677,16 +695,26 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		/**
 		 * Attempts to retrieve task data for all site tasks.
 		 *
+		 * @since 3.1.0 Marked $opt_fields param as deprecated.
 		 * @since 1.0.0
 		 *
-		 * @param string $opt_fields A csv of task fields to retrieve.
+		 * @param string $opt_fields_deprecated Deprecated.
 		 * @return \stdClass[] Task data objects.
 		 *
 		 * @throws \Exception Authentication may fail when first loading the client
 		 * or requests could fail due to request limits or server issues.
 		 */
-		public static function maybe_get_all_site_tasks( string $opt_fields ) : array {
+		public static function maybe_get_all_site_tasks( string $opt_fields_deprecated = '' ) : array {
 
+			if ( ! empty( $opt_fields_deprecated ) ) {
+				_deprecated_argument(
+					__FUNCTION__,
+					'3.1.0',
+					'$opt_fields is now a member constant, ' . __CLASS__ . '::TASK_OPT_FIELDS'
+				);
+			}
+
+			// Load client to ensure current user is authenticated and set.
 			$asana = self::get_client();
 
 			$tasks = [];
@@ -697,7 +725,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 			}
 
 			$params = [
-				'opt_fields' => $opt_fields,
+				'opt_fields' => self::TASK_OPT_FIELDS,
 			];
 
 			$options = [
@@ -712,7 +740,8 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 			$site_tasks = $asana->tasks->findByTag( $site_tag_gid, $params, $options );
 			$all_tasks = [];
 			foreach ( $site_tasks as $task ) {
-				$all_tasks[] = $task;
+				$task->action_link = HTML_Builder::get_task_action_link( $task->gid );
+				$all_tasks[ $task->gid ] = $task;
 			}
 
 			return $all_tasks;
@@ -721,19 +750,28 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		/**
 		 * Sends a batch request to tag and comment on a task in Asana.
 		 *
+		 * @since 3.1.0 Marked $opt_fields param as deprecated.
 		 * @since 1.0.0
 		 *
 		 * @param string $task_gid The task to act on.
 		 * @param string $tag_gid The tag to add.
 		 * @param string $comment The comment text.
-		 * @param string $opt_fields Optional. A csv of task fields to retrieve.
+		 * @param string $opt_fields_deprecated Deprecated.
 		 * @return \stdClass[] An array of response objects. Instance members
 		 * include 'body', 'status_code', and 'headers'.
 		 *
 		 * @throws \Exception Authentication may fail when first loading the client
 		 * or requests could fail due to request limits or server issues.
 		 */
-		public static function tag_and_comment( string $task_gid, string $tag_gid, string $comment, string $opt_fields = '' ) : array {
+		public static function tag_and_comment( string $task_gid, string $tag_gid, string $comment, string $opt_fields_deprecated = '' ) : array {
+
+			if ( ! empty( $opt_fields_deprecated ) ) {
+				_deprecated_argument(
+					__FUNCTION__,
+					'3.1.0',
+					'$opt_fields is now a member constant, ' . __CLASS__ . '::TASK_OPT_FIELDS'
+				);
+			}
 
 			$asana = self::get_client();
 
@@ -741,7 +779,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 			$tag_gid = Options::sanitize( 'gid', $tag_gid );
 			$comment = Options::sanitize( 'string', $comment );
 
-			$opt_fields = explode( ',', $opt_fields );
+			$opt_fields = explode( ',', self::TASK_OPT_FIELDS );
 
 			$data = [
 				'actions' => [
@@ -1018,6 +1056,23 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		}
 
 		/**
+		 * Deletes a task in Asana.
+		 *
+		 * @since 3.1.0
+		 *
+		 * @param string $task_gid The task gid to delete.
+		 *
+		 * @throws \Exception Authentication may fail when first loading the client
+		 * or requests could fail due to request limits or server issues.
+		 */
+		public static function delete_task( string $task_gid ) {
+			// Request the deletion in Asana.
+			$asana = self::get_client();
+			$asana->tasks->delete( $task_gid );
+			// @TODO - Update the cache.
+		}
+
+		/**
 		 * Creates a task in Asana and optionally pins it to a WordPress post.
 		 *
 		 * @since 1.1.0
@@ -1131,6 +1186,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 				}
 			}
 
+			// @TODO - Update cache to contain new task.
 			return $task;
 		}//end create_task()
 	}//end class