+++ src/includes/class-options.php
@@ -609,6 +609,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Options' ) ) {
 					FROM $wpdb->postmeta
 					WHERE meta_key = %s
 						AND meta_value = %s
+					LIMIT 1
 					",
 					$key,
 					$value