+++ changelog.md
@@ -1,3 +1,23 @@
+### 3.1.0 - 2021-12-25
+
+#### Deprecated
+
+- `$opt_fields` arguments in all relevant `Asana_Interface` methods, using new `Asana_Interface::TASK_OPT_FIELDS` member constant instead for consistency.
+
+#### Added
+
+- Many new ReactJS components to compose the new Tasks dashboard widget.
+
+#### Changed
+
+- Tasks dashboard widget has been converted to ReactJS, offering a better experience.
+- Tasks dashboard widget style overhaul, offering a better experience.
+
+#### Fixed
+
+- `Asana_Interface::$wp_user_id` would not be properly set to the current user's ID when loading the Asana client where user ID is `0`, the default value.
+- Newlines not rendering in the dashboard widget's task descriptions.
+
 ### 3.0.1 - 2021-11-09
 
 #### Fixed